// libs/attendance/data-access/src/lib/generated/models/occupancy-day-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { OfficeOccupancyResponse } from '../models/office-occupancy-response';
import { RoomOccupancyResponse } from '../models/room-occupancy-response';
export interface OccupancyDayResponse {
  date: string;
  office: OfficeOccupancyResponse;
  rooms: Array<RoomOccupancyResponse>;
}